hr flow done2

+++ src/main/java/org/example/sijalsystem/Service/RequestInterviewService.java
@@ -2,16 +2,15 @@ package org.example.sijalsystem.Service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.sijalsystem.API.APIException;
-import org.example.sijalsystem.Model.Customer;
-import org.example.sijalsystem.Model.HR;
-import org.example.sijalsystem.Model.InterviewWithHR;
-import org.example.sijalsystem.Model.RequestInterview;
+import org.example.sijalsystem.Model.*;
 import org.example.sijalsystem.Repository.CustomerRepository;
 import org.example.sijalsystem.Repository.HrRepository;
 import org.example.sijalsystem.Repository.InterviewWithHrRepository;
 import org.example.sijalsystem.Repository.RequestInterviewRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.util.Comparator;
 import java.util.List;
 
 @Service
@@ -32,45 +31,71 @@ public class RequestInterviewService {
     public void sendRequestInterview(Integer customer_id, Integer hr_id, RequestInterview requestInterview) {
         HR hr = hrRepository.findHRById(hr_id);
         Customer customer = customerRepository.findCustomerByUser_Id(customer_id);
+
         if (hr == null || customer == null) {
             throw new APIException("HR or Customer not found");
         }
+
+        Subscription lastSubscription = customer.getSubscriptionSet().stream()
+                .max(Comparator.comparing(Subscription::getId))
+                .orElse(null);
+
+        if (lastSubscription != null && lastSubscription.getEndDate().isBefore(LocalDate.now())) {
+            throw new APIException("Subscription is expired");
+        }
+
+        if (!customer.getRequestInterviewSet().isEmpty()) {
+            throw new APIException("You can only send an interview request once");
+        }
+
         requestInterview.setStatus("PENDING");
         requestInterview.setHr(hr);
         requestInterview.setCustomer(customer);
         requestInterviewRepository.save(requestInterview);
 
-        String subject = "طلب مقابلة من " + customer.getUser().getName() + " 📩";
-
-        String body =
+        sendMailService.sendMessage(
+                hr.getUser().getEmail(),
+                "طلب مقابلة من " + customer.getUser().getName() + " 📩",
                 "مرحبًا " + hr.getUser().getName() + ",\n\n" +
                         "لقد قام " + customer.getUser().getName() + " بطلب مقابلة معكم.\n\n" +
                         "رسالة العميل:\n" +
                         "\"" + requestInterview.getMessage() + "\"\n\n" +
                         "موعد المقابلة المقترح: " + requestInterview.getStartTime() + "\n\n" +
-                        "يرجى مراجعة الطلب والرد على العميل بالموافقة أو تحديد موعد آخر.\n\n" +
                         "مع تحياتنا،\n" +
-                        "نظام إدارة المقابلات";
-
-        sendMailService.sendMessage(customer.getUser().getEmail(), subject, body);
+                        "نظام إدارة المقابلات"
+        );
     }
 
-    public void updateRequestInterview(Integer requestInterview_id, RequestInterview requestInterview) {
-        RequestInterview oldRequestInterview = requestInterviewRepository.findRequestInterviewById(requestInterview_id);
 
-        if (oldRequestInterview == null) {
-            throw new APIException("Request interview not found");
+
+    public void updateRequestInterview(Integer customer_id,Integer requestInterview_id, RequestInterview requestInterview) {
+        RequestInterview oldRequestInterview = requestInterviewRepository.findRequestInterviewById(requestInterview_id);
+        Customer customer = customerRepository.findCustomerById(customer_id);
+
+        if (oldRequestInterview == null || customer == null) {
+            throw new APIException("Request interview or customer not found");
+        }
+        if(!requestInterview.getCustomer().getId().equals(customer.getId())){
+            throw new APIException("Customer not authorized to delete this request");
         }
         oldRequestInterview.setMessage(requestInterview.getMessage());
         oldRequestInterview.setStartTime(requestInterview.getStartTime());
         requestInterviewRepository.save(oldRequestInterview);
     }
 
-    public void deleteRequestInterview(Integer requestInterview_id) {
+    public void deleteRequestInterview(Integer customer_id ,Integer requestInterview_id) {
         RequestInterview requestInterview = requestInterviewRepository.findRequestInterviewById(requestInterview_id);
-        if (requestInterview == null) {
-            throw new APIException("Request interview not found");
+        Customer customer = customerRepository.findCustomerById(customer_id);
+        if (requestInterview == null || customer == null) {
+            throw new APIException("Request interview or customer not found");
         }
+        if(requestInterview.getStatus().equalsIgnoreCase("APPROVE")){
+            throw new APIException("You can't delete approved interview request");
+        }
+        if(!requestInterview.getCustomer().getId().equals(customer.getId())){
+            throw new APIException("Customer not authorized to delete this request");
+        }
+
         requestInterviewRepository.delete(requestInterview);
     }
 
@@ -121,6 +146,8 @@ public class RequestInterviewService {
                         "نود إشعاركم بأنه تم قبول طلب الاجتماع بنجاح، وتم تأكيد الموعد المحدد مع العميل.\n\n" +
                         "🗓 موعد الاجتماع:\n" +
                         requestInterview.getStartTime() + "\n\n" +
+                        " ID الاجتماع :\n" +
+                        interviewWithHR.getId() +
                         "🔗 رابط الاجتماع:\n" +
                         meetingLink + "\n\n" +
                         "يرجى التأكد من الانضمام إلى الاجتماع في الوقت المحدد.\n" +
